Split code for protocol base class and ready-to-use core protocols
